+++ packages/cozy-harvest-lib/src/components/TriggerManager.jsx
@@ -246,12 +246,13 @@ export class DumbTriggerManager extends Component {
   }
 
   showCiphersList(ciphers) {
-    this.setState(prevState => {
-      return {
-        step: prevState.step === 'accountForm' ? 'accountForm' : 'ciphersList',
-        ciphers: ciphers ? ciphers : prevState.ciphers
-      }
-    })
+    const newState = { step: 'ciphersList' }
+
+    if (ciphers) {
+      newState.ciphers = ciphers
+    }
+
+    this.setState(newState)
   }
 
   async handleVaultUnlock() {
@@ -275,7 +276,11 @@ export class DumbTriggerManager extends Component {
       if (ciphers.length === 0) {
         this.showAccountForm()
       } else {
-        this.showCiphersList(ciphers)
+        if (this.state.step === 'accountForm') {
+          this.setState({ ciphers })
+        } else {
+          this.showCiphersList(ciphers)
+        }
       }
     } catch (err) {
       logger.error(